[GEM] Bump `PryDoc` to `0.4.5`
Bump it, because I also changed the version of YARD we depend on. I
changed it from "~> 0.8.1" to strict "0.8.3", because of the bug in YARD
"0.8.4", which is the latest version for the time being.

More info: https://github.com/pry/pry/issues/840

Also, I converted EOLs in "Rakefile" and "version.rb" to UNIX (the were
using DOS ones, for some strange reason).